Update host.go
fix int16 sast error

+++ internal/routers/host/host.go
@@ -161,13 +161,13 @@ func Remove(c *gin.Context) {
 		c.String(http.StatusOK, result)
 		return
 	}
-	if id > math.MaxInt16 || id<math.MinInt16 {
+	if id > math.MaxInt16 || id < math.MinInt16 {
 		result=json.CommonFailure(i18n.T(c,"param_error"),fmt.Errorf("host id %d over int16"),id)
 		c.String(http.StatusOK,result)
 		return
 	}
 	taskHostModel := new(models.TaskHost)
-	exist, err := taskHostModel.HostIdExist(int16(id))
+	exist, err := taskHostModel.HostIdExist(id)
 	if err != nil {
 		result = json.CommonFailure(i18n.T(c, "operation_failed"), err)
 		c.String(http.StatusOK, result)
